feat: extract createEncryptedSubmissionDto util fn
