+++ src/app/modules/submission/encrypt-submission/encrypt-submission.utils.ts
@@ -1,6 +1,8 @@
 import { StatusCodes } from 'http-status-codes'
+import moment from 'moment-timezone'
 
 import { createLoggerWithLabel } from '../../../../config/logger'
+import { EncryptedSubmissionDto, SubmissionData } from '../../../../types'
 import { MapRouteError } from '../../../../types/routing'
 import {
   CaptchaConnectionError,
@@ -167,5 +169,24 @@ export const mapRouteError: MapRouteError = (
         statusCode: StatusCodes.INTERNAL_SERVER_ERROR,
         errorMessage: 'Something went wrong. Please try again.',
       }
+  }
+}
+
+/**
+ * Creates and returns an EncryptedSubmissionDto object from submissionData and
+ * attachment presigned urls.
+ */
+export const createEncryptedSubmissionDto = (
+  submissionData: SubmissionData,
+  attachmentPresignedUrls: Record<string, string>,
+): EncryptedSubmissionDto => {
+  return {
+    refNo: submissionData._id,
+    submissionTime: moment(submissionData.created)
+      .tz('Asia/Singapore')
+      .format('ddd, D MMM YYYY, hh:mm:ss A'),
+    content: submissionData.encryptedContent,
+    verified: submissionData.verifiedContent,
+    attachmentMetadata: attachmentPresignedUrls,
   }
 }
